backend/serializers.py
from rest_framework import serializers
from backend.models import Record
from django.contrib.auth.models import User

class RecordSerializer(serializers.ModelSerializer):
    class Meta:
        model = Record
        fields = '__all__'

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'
