requirements.txt
asgiref==3.6.0
Django==4.2
django-bootstrap5==23.1
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.3
numpy==1.24.2
pandas==2.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.9
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
